Update crane-40.3.2.xml
Changed Class Role Topic to Theme, added selectional restriction [+information & +emotional_state] to Theme. Susan and I were just noting that Curtsey and Crane were not in alignment with the other transfer_info classes' use of Theme as opposed to Topic when the information transferred is an emotional state of the Agent. Overstate has the opposite problem.